requirements.txt
Django==3.2.9
djangorestframework==3.13.1
requests==2.27.1
psycopg2-binary==2.9.3
